// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** Generic Array Scalar Type */
  BlockAttributesArray: { input: any; output: any; }
  /** Generic Object Scalar Type */
  BlockAttributesObject: { input: any; output: any; }
};
// ...
export type AcfFieldGroup = {
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};

/** Fields associated with an ACF Field Group */
export type AcfFieldGroupFields = {
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};
// ...
export type AcfLink = {
  __typename?: 'AcfLink';
  /** The target of the link (_blank, etc) */
  target?: Maybe<Scalars['String']['output']>;
  /** The title of the link */
  title?: Maybe<Scalars['String']['output']>;
  /** The url of the link */
  url?: Maybe<Scalars['String']['output']>;
};
// ...
export type BlocksWhoWeAre = AcfFieldGroup & AcfFieldGroupFields & BlocksWhoWeAre_Fields & {
  __typename?: 'BlocksWhoWeAre';
  /** Field of the &quot;group&quot; Field Type added to the schema as part of the &quot;BlocksWhoWeAre&quot; Field Group */
  ctaWwa?: Maybe<BlocksWhoWeAreCtaWwa>;
  /** Field of the &quot;textarea&quot; Field Type added to the schema as part of the &quot;BlocksWhoWeAre&quot; Field Group */
  descriptionWwa?: Maybe<Scalars['String']['output']>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;BlocksWhoWeAre&quot; Field Group */
  subtitleWwa?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;BlocksWhoWeAre&quot; Field Group */
  titleWwa?: Maybe<Scalars['String']['output']>;
};

/** The &quot;BlocksWhoWeAreCtaWwa&quot; Field Group. Added to the Schema by &quot;WPGraphQL for ACF&quot;. */
export type BlocksWhoWeAreCtaWwa = AcfFieldGroup & AcfFieldGroupFields & BlocksWhoWeAreCtaWwa_Fields & {
  __typename?: 'BlocksWhoWeAreCtaWwa';
  /** Field of the &quot;link&quot; Field Type added to the schema as part of the &quot;BlocksWhoWeAreCtaWwa&quot; Field Group */
  ctaWwaLink?: Maybe<AcfLink>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;BlocksWhoWeAreCtaWwa&quot; Field Group */
  ctaWwaTitle?: Maybe<Scalars['String']['output']>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};

/** Interface representing fields of the ACF &quot;BlocksWhoWeAreCtaWwa&quot; Field Group */
export type BlocksWhoWeAreCtaWwa_Fields = {
  /** Field of the &quot;link&quot; Field Type added to the schema as part of the &quot;BlocksWhoWeAreCtaWwa&quot; Field Group */
  ctaWwaLink?: Maybe<AcfLink>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;BlocksWhoWeAreCtaWwa&quot; Field Group */
  ctaWwaTitle?: Maybe<Scalars['String']['output']>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};

/** Interface representing fields of the ACF &quot;BlocksWhoWeAre&quot; Field Group */
export type BlocksWhoWeAre_Fields = {
  /** Field of the &quot;group&quot; Field Type added to the schema as part of the &quot;BlocksWhoWeAre&quot; Field Group */
  ctaWwa?: Maybe<BlocksWhoWeAreCtaWwa>;
  /** Field of the &quot;textarea&quot; Field Type added to the schema as part of the &quot;BlocksWhoWeAre&quot; Field Group */
  descriptionWwa?: Maybe<Scalars['String']['output']>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;BlocksWhoWeAre&quot; Field Group */
  subtitleWwa?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;BlocksWhoWeAre&quot; Field Group */
  titleWwa?: Maybe<Scalars['String']['output']>;
};